+++ dmenu_hotkeys/parsers.py
@@ -1,3 +1,6 @@
+from dmenu_hotkeys.constants import PARSERS
+
+
 class BaseConfigParser(object):
     def parse_hotkey(self, line):
         """
@@ -28,3 +31,15 @@ class OpenBoxConfigParser(BaseConfigParser):
         :return: string
         """
         return line.split("\"")[1]  # todo add xml parser
+
+
+def get_parser(app_name):
+    """
+    Return parser based on app name.
+    :param app_name: String
+    :return: parser object
+    """
+    app_parser_name = PARSERS[app_name]
+    parser = [subcls for subcls in BaseConfigParser.__subclasses__() if
+              subcls.__name__ == app_parser_name][0]
+    return parser()  # return parser object instead of class