+++ HISTORY.rst
@@ -4,6 +4,9 @@ History
 
 Unreleased
 ------------------
+* Added config as a singleton.
+* Added conf path passing from console argument.
+* Added testing conf and application fixtures.
 * Added copy-config command to copying config files to the user home dir.
 * Added menu and apps installation validation.